requirements.txt
cryptography
boto3
pytz
hidapi
pyasn1
requests_aws4auth